safetychecker/requirements.txt
openai>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0
python-dotenv>=0.19.0